services/ml_inference_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
# For ML:
pillow==10.0.0
rawpy==0.19.0
# Torch and related libraries - versioning is critical here
# For CUDA 11.8 based PyTorch (common):
torch==2.1.0
transformers==4.35.2
# For CLIP, from openai
clip @ git+https://github.com/openai/CLIP.git
# Add other dependencies like transformers, torch, PIL for ML models 